Moving V3Global function definitions to their own files
