+++ src/Makefile_obj.in
@@ -193,6 +193,7 @@ RAW_OBJS = \
 	V3FileLine.o \
 	V3Gate.o \
 	V3GenClk.o \
+	V3Global.o \
 	V3Graph.o \
 	V3GraphAlg.o \
 	V3GraphAcyc.o \
